+++ ft_memchr.c
@@ -6,13 +6,15 @@
 /*   By: gpires-c <gpires-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/26 19:10:17 by gpires-c          #+#    #+#             */
-/*   Updated: 2026/05/27 20:38:05 by gpires-c         ###   ########.fr       */
+/*   Updated: 2026/05/29 19:43:19 by gpires-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*p;
 
 	i = 0;
